requirements.txt
pytesseract==0.3.4
opencv-python==4.2.0.34
pandas==0.25.3
Pillow==2.2.2
